fix migrations

+++ db/migrate/20170201141501_remove_emotions.rb
@@ -1,8 +1,8 @@
 class RemoveEmotions < ActiveRecord::Migration[5.0]
   def change
-    drop_table :happies
-    drop_table :sads
-    drop_table :likes
+    drop_table :happies unless table_exists? :happies
+    drop_table :sads unless table_exists? :sads
+    drop_table :likes unless table_exists? :likes
     remove_column :posts, :happies_count
     remove_column :posts, :likes_count
     remove_column :posts, :sads_count
